Added Gemfile.lock to .gitignore
The purpose of this Gemfile is to give us current
master branches where we need them. Thus it makes
no sense to lock all devs to the same versions.

It's advised to run bundle install before working
on something to guarantee testing against current
code.

Since we need to optionally depend on different
do_xxx gems for testing, it cannot be guaranteed
that all developers will have the same do_xxx
gems installed on their machine. The question on
exactly which gems to include in Gemfile.lock
would arise.

For all this reasons, I think it's best to just
not have it in the repository in the first place.