Extracted DataObjects::Schema::Collection (soon Set)
Collection is the base class used for

  DataObjects::Schema::Tables
  DataObjects::Schema::Columns
  DataObjects::Schema::OrderedSet

The Collection currently uses a (hardcoded)
Set class as its internal storage, since all
so far discovered collection classes need
Set semantics anyway. This will change very
soon and Collection will probably be renamed
to

  DataObjects::Schema::Set

and an Array will be used as container.

Collection now has a set of shared specs
that are used by the concrete subclasses.