Unified unique constraint/index manipulation API draft
I think unique indexes can be considered to be an
implementation detail for the greater concept of
enforcing unique constraints.

do-schema should properly construct unique indexes
when a unique constraint is specified and when the
underlying database engine doesn't do so itself, if
that's even the case for any of the engines we want
to support.

I'm still not entirely sure if having a UniqueIndex
class representing that concept, is necessary to
have in do-schema in the first place.